커뮤니티 게시글 수정 / alert와 confirm 구분

+++ src/componets/mypage/components/community.jsx
@@ -20,10 +20,11 @@ const Community = (props) => {
   };
 
   const delBoard = (id) => {
-    alert("정말 삭제하시겠습니까?");
-    deleteCommunity(id).then(() =>
-      fetchByUser().then((res) => setBoard(res.data))
-    );
+    if (confirm("정말 삭제하시겠습니까?") === true) {
+      deleteCommunity(id).then(() =>
+        fetchByUser().then((res) => setBoard(res.data))
+      );
+    }
   };
 
   const goUpdate = (id) => {
